Organize test stubs
