+++ tests/stubs/ultralytics.py
@@ -1,0 +1,6 @@
+class YOLO:
+    def __init__(self, *args, **kwargs):
+        pass
+
+    def __call__(self, *args, **kwargs):
+        return []
