+++ tests/test_annotation_pipeline_full.py
@@ -1,5 +1,4 @@
 import sys
-import types
 from pathlib import Path
 from unittest import mock
 
@@ -7,8 +6,7 @@ import numpy as np
 import yaml
 import shutil
 
-# Ensure the annotation module can be imported without ultralytics installed
-sys.modules['ultralytics'] = types.SimpleNamespace(YOLO=lambda *a, **k: None)
+from tests.stubs import ultralytics  # noqa: F401
 
 MODULE_PATH = Path(__file__).resolve().parents[1] / 'pretraining' / 'annotation'
 sys.path.append(str(MODULE_PATH))
