+++ tests/test_google_drive_authenticate.py
@@ -1,7 +1,6 @@
 import json
 import os
 import sys
-import types
 import base64
 from pathlib import Path
 
@@ -10,40 +9,7 @@ import pytest
 MODULE_DIR = Path(__file__).resolve().parents[1] / 'code' / 'modules'
 sys.path.append(str(MODULE_DIR))
 
-
-class DummyCreds:
-    pass
-
-
-def setup_google_modules(monkeypatch, captured_info):
-    import google.oauth2.service_account as sa_mod
-    import googleapiclient.discovery as disc_mod
-    import googleapiclient.http as http_mod
-
-    class DummyMediaFileUpload:
-        def __init__(self, filename, resumable=False):
-            self.filename = filename
-
-    class DummyMediaIoBaseDownload:
-        def __init__(self, fh, request):
-            self.fh = fh
-            self.request = request
-            self.done = False
-
-        def next_chunk(self):
-            if not self.done:
-                self.fh.write(self.request.content)
-                self.done = True
-            return None, self.done
-
-    def fake_from_info(info, scopes=None):
-        captured_info['info'] = info
-        return DummyCreds()
-
-    monkeypatch.setattr(sa_mod.Credentials, 'from_service_account_info', fake_from_info, raising=False)
-    monkeypatch.setattr(disc_mod, 'build', lambda *a, **k: 'service', raising=False)
-    monkeypatch.setattr(http_mod, 'MediaFileUpload', DummyMediaFileUpload, raising=False)
-    monkeypatch.setattr(http_mod, 'MediaIoBaseDownload', DummyMediaIoBaseDownload, raising=False)
+from tests.stubs.google_api import setup_google_modules
 
 
 def test_authenticate_accepts_file_path(tmp_path, monkeypatch):
